+++ CHANGELOG.md
@@ -8,12 +8,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- `GeoSeries`in GeometryManager.
-- Bugs when managing multiple indices with accessor `xr.ed`.
+- `GeoSeries` supported in GeometryManager.
+- `ed.sel_nearest_dates` accessor avoid duplicated times.
+- Issue when managing multiple indices with accessor `xr.ed`.
+- Issue when same datetime when rescaling dataset.
+
+### Added
+
+- `mode` for zonal stats in `operations.reducers`.
+
+### Changed
+
+- `ag_cloud_mask_items` queries items per batch.
 
 ### Removed
 
-- `mode` for zonal stats `operations.reducers`.
+- Typing decorator, expected new typing library.
 
 ## [0.0.13] - 2024-03-06